Add challenge question for multiple true branches
A common misconception with conditionals is learners expecting all true blocks to be executed when in reality only the first true block is executed. This multiple choice question provides a quick check to see if the student has picked this up with multiple elif statements.

This is mentioned in the notes for if/else blocks, but not specifically for elif so is worth reiterating in a challenge.