Added exercise on lists
As this issue (=/=swcarpentry/=/=python-novice-inflammation/issues/84) pointed out, there were no exercises for the lists module, so I wrote this one. It hasn't been tested in a software carpentry setting (that I know of), but it worked well in the Intro to Python class that I TAed. It hints at the idea of accumulators, which I think is valuable to touch on.

It would probably be good to also have an exercise or two that deals with indexing into lists, but I'm having trouble coming up with one that is neither obvious from the examples nor really convoluted.