Update some docstring sections to use numpy standard. (/issues/729)
* Update some docstring sections to use numpy standard.

You will also see that some of the information was incorrect due to
multiple update. The function used to be called `center` and to return
a list instead of a numpy array. So the help and example were incorrect.

The use of the numpy standard is a good practice; as most projects use
it. Even if we don't expect users to remember it; at least they are made
aware that there is standard and can look for it.

This also mirror the documentation style of loadtxt that is seen later
in the lesson.

And finally sphinx can understand numpydoc and generate nice HTML pages.
So all in all I believe this is a nice detail to add to the lesson.

* Use minimal numpy docstring with only example.