Update setup.md
I'm updating these links because the commonmark parser does not recognise them and my intervention did not account for this situation. This is the only place in this or any of the other lessons I have modified where this happens, so it's easier for me to update this than to update my intervention and add tests.